chore(adoption-insights): promote graduate nfs /alpha exports to main entry point
