+++ workspaces/adoption-insights/packages/app-legacy/src/App.tsx
@@ -50,8 +50,10 @@ import { AppRouter, FlatRoutes } from '@backstage/core-app-api';
 import { CatalogGraphPage } from '@backstage/plugin-catalog-graph';
 import { RequirePermission } from '@backstage/plugin-permission-react';
 import { catalogEntityCreatePermission } from '@backstage/plugin-catalog-common/alpha';
-import { AdoptionInsightsPage } from '@red-hat-developer-hub/backstage-plugin-adoption-insights';
-import { adoptionInsightsTranslations } from '@red-hat-developer-hub/backstage-plugin-adoption-insights/alpha';
+import {
+  AdoptionInsightsPage,
+  adoptionInsightsTranslations,
+} from '@red-hat-developer-hub/backstage-plugin-adoption-insights/legacy';
 import { githubAuthApiRef } from '@backstage/core-plugin-api';
 
 const app = createApp({
