+++ workspaces/adoption-insights/packages/app/knip-report.md
@@ -1,0 +1,20 @@
+# Knip report
+
+## Unused dependencies (4)
+
+| Name                             | Location          | Severity |
+| :------------------------------- | :---------------- | :------- |
+| @backstage/plugin-app-visualizer | package.json:29:6 | error    |
+| @backstage/integration-react     | package.json:27:6 | error    |
+| @backstage/plugin-search         | package.json:31:6 | error    |
+| react-router                     | package.json:41:6 | error    |
+
+## Unused devDependencies (5)
+
+| Name                           | Location          | Severity |
+| :----------------------------- | :---------------- | :------- |
+| @backstage/frontend-test-utils | package.json:45:6 | error    |
+| @testing-library/user-event    | package.json:50:6 | error    |
+| @testing-library/dom           | package.json:47:6 | error    |
+| @playwright/test               | package.json:46:6 | error    |
+| cross-env                      | package.json:52:6 | error    |
