+++ workspaces/adoption-insights/packages/backend/knip-report.md
@@ -1,15 +1,16 @@
 # Knip report
 
-## Unused dependencies (9)
+## Unused dependencies (10)
 
-| Name                                  | Location          | Severity |
-| :------------------------------------ | :---------------- | :------- |
-| @backstage/plugin-search-backend-node | package.json:46:6 | error    |
-| @backstage/plugin-permission-common   | package.json:37:6 | error    |
-| @backstage/plugin-permission-node     | package.json:38:6 | error    |
-| @backstage/plugin-auth-node           | package.json:30:6 | error    |
-| @backstage/config                     | package.json:25:6 | error    |
-| better-sqlite3                        | package.json:50:6 | error    |
-| node-gyp                              | package.json:51:6 | error    |
-| app                                   | package.json:49:6 | error    |
-| pg                                    | package.json:52:6 | error    |
+| Name                                                             | Location          | Severity |
+| :--------------------------------------------------------------- | :---------------- | :------- |
+| @red-hat-developer-hub/backstage-plugin-adoption-insights-common | package.json:50:6 | error    |
+| @backstage/plugin-search-backend-node                            | package.json:47:6 | error    |
+| @backstage/plugin-permission-common                              | package.json:38:6 | error    |
+| @backstage/plugin-permission-node                                | package.json:39:6 | error    |
+| @backstage/backend-plugin-api                                    | package.json:25:6 | error    |
+| @backstage/plugin-auth-node                                      | package.json:31:6 | error    |
+| @backstage/config                                                | package.json:26:6 | error    |
+| node-gyp                                                         | package.json:53:6 | error    |
+| app                                                              | package.json:51:6 | error    |
+| pg                                                               | package.json:54:6 | error    |
