+++ workspaces/adoption-insights/plugins/adoption-insights/package.json
@@ -2,18 +2,22 @@
   "name": "@red-hat-developer-hub/backstage-plugin-adoption-insights",
   "version": "0.9.1",
   "license": "Apache-2.0",
-  "main": "src/index.ts",
-  "types": "src/index.ts",
+  "main": "src/index.tsx",
+  "types": "src/index.tsx",
   "exports": {
-    ".": "./src/index.ts",
-    "./alpha": "./src/alpha.tsx",
+    ".": "./src/index.tsx",
+    "./alpha": "./src/alpha.ts",
+    "./legacy": "./src/legacy.ts",
     "./adoption-insights-translations-module": "./src/adoptionInsightsTranslationsModuleExport.ts",
     "./package.json": "./package.json"
   },
   "typesVersions": {
     "*": {
       "alpha": [
-        "src/alpha.tsx"
+        "src/alpha.ts"
+      ],
+      "legacy": [
+        "src/legacy.ts"
       ],
       "adoption-insights-translations-module": [
         "src/adoptionInsightsTranslationsModuleExport.ts"
@@ -24,11 +28,7 @@
     }
   },
   "publishConfig": {
-    "access": "public",
-    "alpha": {
-      "types": "dist/alpha.d.ts",
-      "default": "dist/alpha.esm.js"
-    }
+    "access": "public"
   },
   "repository": {
     "type": "git",
@@ -98,8 +98,9 @@
   "scalprum": {
     "name": "red-hat-developer-hub.backstage-plugin-adoption-insights",
     "exposedModules": {
-      "PluginRoot": "./src/index.ts",
-      "Alpha": "./src/alpha.tsx"
+      "PluginRoot": "./src/index.tsx",
+      "Legacy": "./src/legacy.ts",
+      "Alpha": "./src/alpha.ts"
     }
   }
 }
