+++ workspaces/adoption-insights/plugins/adoption-insights/report-adoption-insights-translations-module.api.md
@@ -5,7 +5,7 @@
 ```ts
 import { FrontendModule } from '@backstage/frontend-plugin-api';
 
-// @alpha
+// @public
 const adoptionInsightsTranslationsModule: FrontendModule;
 export default adoptionInsightsTranslationsModule;
