+++ workspaces/adoption-insights/plugins/adoption-insights/src/legacy.ts
@@ -13,24 +13,24 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-import {
-  analyticsApiRef,
-  configApiRef,
-  createApiFactory,
-  identityApiRef,
-} from '@backstage/core-plugin-api';
-import { AdoptionInsightsAnalyticsApi } from './AdoptionInsightsAnalyticsApi';
 
 /**
- * API Factory for Adoption Insights Analytics API
+ * Legacy frontend system API surface for the Adoption Insights plugin.
  *
- * @public
+ * @packageDocumentation
  */
-export const AdoptionInsightsAnalyticsApiFactory = createApiFactory({
-  api: analyticsApiRef,
-  deps: { configApi: configApiRef, identityApi: identityApiRef },
-  factory: ({ configApi, identityApi }) =>
-    AdoptionInsightsAnalyticsApi.fromConfig(configApi, {
-      identityApi,
-    }),
+
+import { unstable_ClassNameGenerator as ClassNameGenerator } from '@mui/material/className';
+
+ClassNameGenerator.configure(componentName => {
+  return componentName.startsWith('v5-')
+    ? componentName
+    : `v5-${componentName}`;
 });
+
+export * from './legacy/plugin';
+
+export {
+  adoptionInsightsTranslations,
+  adoptionInsightsTranslationRef,
+} from './translations';
