+++ workspaces/adoption-insights/plugins/adoption-insights/src/translations/index.ts
@@ -20,7 +20,7 @@ import { adoptionInsightsTranslationRef } from './ref';
 
 /**
  * The translation resource for the Adoption Insights plugin.
- * @alpha
+ * @public
  */
 export const adoptionInsightsTranslations = createTranslationResource({
   ref: adoptionInsightsTranslationRef,
