+++ workspaces/adoption-insights/plugins/analytics-module-adoption-insights/knip-report.md
@@ -1,20 +1,17 @@
 # Knip report
 
-## Unused dependencies (4)
+## Unused dependencies (1)
 
-| Name               | Location          | Severity |
-| :----------------- | :---------------- | :------- |
-| @material-ui/icons | package.json:38:6 | error    |
-| @material-ui/core  | package.json:37:6 | error    |
-| @material-ui/lab   | package.json:39:6 | error    |
-| react-use          | package.json:40:6 | error    |
+| Name      | Location          | Severity |
+| :-------- | :---------------- | :------- |
+| react-use | package.json:50:6 | error    |
 
 ## Unused devDependencies (5)
 
 | Name                        | Location          | Severity |
 | :-------------------------- | :---------------- | :------- |
-| @testing-library/user-event | package.json:52:6 | error    |
-| @backstage/core-app-api     | package.json:47:6 | error    |
-| @testing-library/react      | package.json:51:6 | error    |
-| @backstage/test-utils       | package.json:49:6 | error    |
-| msw                         | package.json:53:6 | error    |
+| @testing-library/user-event | package.json:62:6 | error    |
+| @backstage/core-app-api     | package.json:57:6 | error    |
+| @testing-library/react      | package.json:61:6 | error    |
+| @backstage/test-utils       | package.json:59:6 | error    |
+| msw                         | package.json:63:6 | error    |
