+++ workspaces/adoption-insights/plugins/analytics-module-adoption-insights/report.api.md
@@ -3,38 +3,35 @@
 > Do not edit this file. It is a report generated by [API Extractor](https://api-extractor.com/).
 
 ```ts
-import { AnalyticsApi } from '@backstage/core-plugin-api';
-import { AnalyticsEvent } from '@backstage/frontend-plugin-api';
-import { AnalyticsEvent as AnalyticsEvent_2 } from '@backstage/core-plugin-api';
-import { AnalyticsImplementation } from '@backstage/frontend-plugin-api';
-import { ApiFactory } from '@backstage/frontend-plugin-api';
-import { Config } from '@backstage/config';
-import { ConfigApi } from '@backstage/core-plugin-api';
-import { IdentityApi } from '@backstage/core-plugin-api';
-import { IdentityApi as IdentityApi_2 } from '@backstage/frontend-plugin-api';
+import { AnalyticsImplementationFactory } from '@backstage/frontend-plugin-api';
+import { ExtensionBlueprintParams } from '@backstage/frontend-plugin-api';
+import { ExtensionDataRef } from '@backstage/frontend-plugin-api';
+import { OverridableExtensionDefinition } from '@backstage/frontend-plugin-api';
+import { OverridableFrontendPlugin } from '@backstage/frontend-plugin-api';
 
 // @public
-export class AdoptionInsightsAnalyticsApi
-  implements AnalyticsApi, AnalyticsImplementation
-{
-  captureEvent(event: AnalyticsEvent | AnalyticsEvent_2): Promise<void>;
-  static fromConfig(
-    config: ConfigApi,
-    options: {
-      identityApi: IdentityApi;
-    },
-  ): AdoptionInsightsAnalyticsApi;
-}
-
-// @public
-export const AdoptionInsightsAnalyticsApiFactory: ApiFactory<
-  AnalyticsApi,
-  AdoptionInsightsAnalyticsApi,
+const _default: OverridableFrontendPlugin<
+  {},
+  {},
   {
-    configApi: Config;
-    identityApi: IdentityApi_2;
+    'analytics:analytics-module-adoption-insights/analytics-module-adoption-insights': OverridableExtensionDefinition<{
+      kind: 'analytics';
+      name: 'analytics-module-adoption-insights';
+      config: {};
+      configInput: {};
+      output: ExtensionDataRef<
+        AnalyticsImplementationFactory<{}>,
+        'core.analytics.factory',
+        {}
+      >;
+      inputs: {};
+      params: <TDeps extends { [name in string]: unknown }>(
+        params: AnalyticsImplementationFactory<TDeps>,
+      ) => ExtensionBlueprintParams<AnalyticsImplementationFactory<{}>>;
+    }>;
   }
 >;
+export default _default;
 
 // (No @packageDocumentation comment for this package)
 ```
