--- workspaces/adoption-insights/plugins/analytics-module-adoption-insights/src/alpha.ts
@@ -1,41 +1,0 @@
-/*
- * Copyright Red Hat, Inc.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-import {
-  AnalyticsImplementationBlueprint,
-  configApiRef,
-  createFrontendPlugin,
-  identityApiRef,
-} from '@backstage/frontend-plugin-api';
-import { AdoptionInsightsAnalyticsApi } from './api/AdoptionInsightsAnalyticsApi';
-
-const adoptionInsightsAnalyticsImplementation =
-  AnalyticsImplementationBlueprint.make({
-    name: 'analytics-module-adoption-insights',
-    params: defineParams =>
-      defineParams({
-        deps: { configApi: configApiRef, identityApi: identityApiRef },
-        factory: ({ configApi, identityApi }) =>
-          AdoptionInsightsAnalyticsApi.fromConfig(configApi, { identityApi }),
-      }),
-  });
-
-/**
- * @alpha
- */
-export default createFrontendPlugin({
-  pluginId: 'analytics-module-adoption-insights',
-  extensions: [adoptionInsightsAnalyticsImplementation],
-});
