+++ workspaces/adoption-insights/plugins/analytics-module-adoption-insights/src/nfsExports.test.ts
@@ -13,4 +13,15 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-export { adoptionInsightsTranslationsModule as default } from './index';
+import plugin from './index';
+
+describe('analytics-module-adoption-insights NFS exports', () => {
+  it('should export a valid frontend plugin', () => {
+    expect(plugin).toBeDefined();
+    expect(plugin.$$type).toBe('@backstage/FrontendPlugin');
+  });
+
+  it('should have the correct plugin id', () => {
+    expect(plugin.id).toBe('analytics-module-adoption-insights');
+  });
+});
